Salvando as informações no banco de dados MYSQL e add update no AlunoController

+++ src/main/java/com/bruno/controleProjetos/domain/Aluno.java
@@ -16,7 +16,7 @@ public class Aluno implements Serializable{
 		
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Integer id; 
+	private Long id; 
 	private String nome;
 	private String matricula;
 	
@@ -28,17 +28,17 @@ public class Aluno implements Serializable{
 	}
 	
 	
-	public Aluno(Integer id, String nome, String matricula) {
+	public Aluno(Long id, String nome, String matricula) {
 		super();
 		this.id = id;
 		this.nome = nome;
 		this.matricula = matricula;
 		
 	}
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
 	public String getNome() {
